Update some functions in vacuum_iso_bubbles, prep for 6 loops.
Fixed a few small bugs in vacuum_iso_bubbles involving factorized graphs,
which we need to use for loop->loop reduction.  Additionally added
functionality to findVacuumRep so that it searches the full basis
for an isomorphic graph, which is relevant when moving around bubbles.

Also built files that are necessary for generating the different 6 loop
towers.